+++ app/assets/javascripts/central.js
@@ -5,6 +5,8 @@ var ProjectView = require('views/project_view');
 var ProjectSearchView = require('views/project_search_view');
 var ProjectVelocityView = require('views/project_velocity_view');
 
+import TourController from 'controllers/tour/TourController';
+
 require('./global_listeners');
 
 var Central = module.exports = {
@@ -34,5 +36,10 @@ var Central = module.exports = {
 
       window.projectView = view;
     });
+
+    setTimeout(() => {
+      const tourController = new TourController();
+      tourController.initialize();
+    }, 200);
   }
 };