+++ app/controllers/registrations_controller.rb
@@ -26,6 +26,29 @@ class RegistrationsController < Devise::RegistrationsController
     end
   end
 
+  def tour
+    @user = User.find(params[:id])
+    if @user.update(user_params)
+      render json: @user, status: :ok
+    else
+      render json: @user.errors, status: :unprocessable_entity
+    end
+  end
+
+  def reset_tour
+    @user = User.find(params[:id])
+    if @user.update(user_params)
+      redirect_to :back, flash: {notice: t('reset_tour_success')}
+    else
+      redirect_to :back, flash: {error: t('reset_tour_fail')}
+    end
+  end
+
+  private
+    def user_params
+      params.require(:user).permit(:finished_tour)
+    end
+
   protected
     def after_inactive_sign_up_path_for(resource)
       new_session_path(resource)